Hybrid
* Moved all files from hybrid to hybrid legacy

+++ Engine/Framework/Internal/SDL3/SDL/Core/Rect.cs
@@ -5,9 +5,9 @@ internal static unsafe partial class SDL
     [StructLayout(LayoutKind.Sequential)]
     public struct Rect
     {
-        public int x;
-        public int y;
-        public int w;
-        public int h;
+        public float x;
+        public float y;
+        public float w;
+        public float h;
     }
 }
